SortableBookshelfCard: Replace SortableCollectionBookCard,; Move SortableBookshelfCardOptions from SortableBookshelfContext
